backend/user_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
python-dotenv==1.0.0
pydantic[email]==2.4.2
motor==3.3.2 